moved some int tests
